+++ src/DependencyGraph/AddressMapping/DenseStrategy.ts
@@ -171,7 +171,7 @@ export class DenseStrategy implements IAddressMappingStrategy {
 
   public* entriesFromColumnsSpan(columnsSpan: ColumnsSpan): IterableIterator<[SimpleCellAddress, CellVertex]> {
     for (let x = columnsSpan.columnStart; x <= columnsSpan.columnEnd; ++x) {
-      for (let y = 0; y < this.height; ++y) {    
+      for (let y = 0; y < this.height; ++y) {
         const vertex = this.getCellVertex(x, y)
         if (vertex) {
           yield [simpleCellAddress(columnsSpan.sheet, x, y), vertex]