+++ src/HyperFormula.ts
@@ -423,7 +423,9 @@ export class HyperFormula implements TypedEmitter {
   /**
    * Registers all functions in a given plugin with optional translations.
    *
-   * Note: This method does not affect the existing HyperFormula instances.
+   * Note: FunctionPlugins must be registered prior to the creation of HyperFormula instances in which they are used.
+   * HyperFormula instances created prior to the registration of a FunctionPlugin are unable to access the FunctionPlugin.
+   * Registering a FunctionPlugin with [[custom-functions]] requires the translations parameter.
    *
    * @param {FunctionPluginDefinition} plugin - plugin class
    * @param {FunctionTranslationsPackage} translations - optional package of function names translations
@@ -3881,7 +3883,7 @@ export class HyperFormula implements TypedEmitter {
       this._emitter.emit(Events.NamedExpressionRemoved, removedNamedExpression.displayName, changes)
       return changes
     } else {
-      return []
+      return [] // codecov note: this does not look possible - removeNamedExpression() will throw if the named expression cannot be found
     }
   }
 
@@ -4056,7 +4058,7 @@ export class HyperFormula implements TypedEmitter {
     }
 
     if (ast.type === AstNodeType.ERROR && !ast.error) {
-      return false
+      return false // codecov note: could not identify a formulaString that would cause this condition
     }
 
     return true